fix(frontend+backend): resolve SSE rendering, asset 404s, and Decimal serialization
- customer/hooks/useChat.ts: buffer partial SSE lines across stream chunks
  to prevent lost/incomplete chat responses in production builds. Clear
  loading state immediately on [DONE] instead of waiting for stream close.
- main.py: add explicit static routes for /favicon.svg and /icons.svg to
  prevent 404s when FastAPI serves the built frontend.
- index.html & admin.html: replace /vite.svg with /favicon.svg.
- models/order.py: change total_amount from float to Decimal to match
  Numeric column type and eliminate PydanticSerializationUnexpectedValue
  warnings.
- .gitignore: add .playwright-cli/, *.yaml, *.yml, and .sisyphus/

+++ app/models/order.py
@@ -1,5 +1,6 @@
 # app/models/order.py
 from datetime import datetime
+from decimal import Decimal
 from enum import Enum
 
 from sqlalchemy import JSON, Column, DateTime, Numeric, String, text
@@ -31,7 +32,7 @@ class Order(SQLModel, table=True):
         default=OrderStatus.PENDING, sa_column=Column(String, index=True, nullable=False)
     )
 
-    total_amount: float = Field(sa_column=Column(Numeric(precision=10, scale=2)))
+    total_amount: Decimal = Field(sa_column=Column(Numeric(precision=10, scale=2)))
     items: list[dict] = Field(default_factory=list, sa_column=Column(JSON))
 
     tracking_number: str | None = Field(default=None, index=True)
